refactor: Split audit_service.py into focused services (SRP)
- Extract AuditEventLogger (185L) for event persistence
- Extract AuditRulesEngine (506L) for validation rules
- Extract AuditOrchestrator (95L) for coordination
- Reduce audit_service.py from 827L to 117L (facade only)
- Maintain 100% backward compatibility
- All 537 tests pass

Fixes #108
