+++ apps/api/services/audit/__init__.py
@@ -1,0 +1,9 @@
+"""
+Audit package - focused services following SRP.
+"""
+
+from .event_logger import AuditEventLogger
+from .rules_engine import AuditRulesEngine
+from .orchestrator import AuditOrchestrator
+
+__all__ = ["AuditEventLogger", "AuditRulesEngine", "AuditOrchestrator"]
